requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
langgraph==0.0.26
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0
google-generativeai==0.3.2